Refactor folder structure of bundle
