+++ src/Sulu/Bundle/CustomUrlBundle/Domain/Event/CustomUrlCreatedEvent.php
@@ -12,7 +12,7 @@
 namespace Sulu\Bundle\CustomUrlBundle\Domain\Event;
 
 use Sulu\Bundle\CustomUrlBundle\Admin\CustomUrlAdmin;
-use Sulu\Bundle\EventLogBundle\Event\DomainEvent;
+use Sulu\Bundle\EventLogBundle\Domain\Event\DomainEvent;
 use Sulu\Component\CustomUrl\Document\CustomUrlDocument;
 
 class CustomUrlCreatedEvent extends DomainEvent
