+++ src/Sulu/Bundle/EventLogBundle/Application/Collector/DomainEventCollectorInterface.php
@@ -9,9 +9,9 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\EventLogBundle\Collector;
+namespace Sulu\Bundle\EventLogBundle\Application\Collector;
 
-use Sulu\Bundle\EventLogBundle\Event\DomainEvent;
+use Sulu\Bundle\EventLogBundle\Domain\Event\DomainEvent;
 
 interface DomainEventCollectorInterface
 {
