+++ src/Sulu/Bundle/EventLogBundle/Application/Subscriber/StoreEventRecordSubscriber.php
@@ -9,10 +9,10 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\EventLogBundle\Subscriber;
+namespace Sulu\Bundle\EventLogBundle\Application\Subscriber;
 
-use Sulu\Bundle\EventLogBundle\Entity\EventRecordRepositoryInterface;
-use Sulu\Bundle\EventLogBundle\Event\DomainEvent;
+use Sulu\Bundle\EventLogBundle\Domain\Event\DomainEvent;
+use Sulu\Bundle\EventLogBundle\Domain\Repository\EventRecordRepositoryInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 class StoreEventRecordSubscriber implements EventSubscriberInterface
