+++ src/Sulu/Bundle/EventLogBundle/Infrastructure/Symfony/DependencyInjection/Configuration.php
@@ -9,10 +9,10 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\EventLogBundle\DependencyInjection;
+namespace Sulu\Bundle\EventLogBundle\Infrastructure\Symfony\DependencyInjection;
 
-use Sulu\Bundle\EventLogBundle\Entity\DoctrineEventRecordRepository;
-use Sulu\Bundle\EventLogBundle\Entity\EventRecord;
+use Sulu\Bundle\EventLogBundle\Domain\Model\EventRecord;
+use Sulu\Bundle\EventLogBundle\Infrastructure\Doctrine\Repository\EventRecordRepository;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
@@ -47,7 +47,7 @@ class Configuration implements ConfigurationInterface
                         ->addDefaultsIfNotSet()
                         ->children()
                             ->scalarNode('model')->defaultValue(EventRecord::class)->end()
-                            ->scalarNode('repository')->defaultValue(DoctrineEventRecordRepository::class)->end()
+                            ->scalarNode('repository')->defaultValue(EventRecordRepository::class)->end()
                         ->end()
                     ->end()
                 ->end()
