+++ src/Sulu/Bundle/EventLogBundle/Resources/config/doctrine/EventRecord.orm.xml
@@ -2,7 +2,7 @@
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-    <mapped-superclass name="Sulu\Bundle\EventLogBundle\Entity\EventRecord" table="el_event_records">
+    <mapped-superclass name="Sulu\Bundle\EventLogBundle\Domain\Model\EventRecord" table="el_event_records">
         <id name="id" type="integer" column="id">
             <generator strategy="AUTO"/>
         </id>
